requirements.txt
appdirs==1.4.4
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
fake-useragent==2.0.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
finnhub-python==2.4.22
fonttools==4.55.3
frozendict==2.4.6
html5lib==1.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
lxml_html_clean==0.4.1
matplotlib==3.9.3
multitasking==0.0.11
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.0
packaging==24.2
pandas==2.2.3
parse==1.20.2
parsel==1.9.1
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==11.1.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
scikit-learn==1.6.0
scipy==1.14.1
Scrapy==2.12.0
seaborn==0.13.2
service-identity==24.2.0
setuptools==75.6.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.3
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vaderSentiment==3.3.2
w3lib==2.2.1
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.2.50
zipp==3.21.0
zope.interface==7.2
